requirements.txt
mysql-connector-python
telebot
babel
python-dateutil